@@ -1,4 +1,4 @@
-import React, { useState, useEffect, useCallback } from "react";
+import React, { useState, useEffect } from "react";
 import throttle from "lodash/fp/throttle";
 
 export interface useScrollSpyParams {
@@ -17,27 +17,23 @@ export default ({
 }: useScrollSpyParams): number | undefined => {
   const [activeSection, setActiveSection] = useState(activeSectionDefault);
 
-  const handle = useCallback(
-    () =>
-      throttle(throttleMs, () => {
-        let currentSectionId = activeSection;
-        for (let i = 0; i < sectionElementRefs.length; i += 1) {
-          const section = sectionElementRefs[i].current;
-          // Needs to be a valid DOM Element
-          if (!section || !(section instanceof Element)) continue;
-          // GetBoundingClientRect returns values relative to viewport
-          if (section.getBoundingClientRect().top + offsetPx < 0) {
-            currentSectionId = i;
-            continue;
-          }
-          // No need to continue loop, if last element has been detected
-          break;
-        }
+  const handle = throttle(throttleMs, () => {
+    let currentSectionId = activeSection;
+    for (let i = 0; i < sectionElementRefs.length; i += 1) {
+      const section = sectionElementRefs[i].current;
+      // Needs to be a valid DOM Element
+      if (!section || !(section instanceof Element)) continue;
+      // GetBoundingClientRect returns values relative to viewport
+      if (section.getBoundingClientRect().top + offsetPx < 0) {
+        currentSectionId = i;
+        continue;
+      }
+      // No need to continue loop, if last element has been detected
+      break;
+    }
 
-        setActiveSection(currentSectionId);
-      }),
-    [activeSection, offsetPx, sectionElementRefs, throttleMs]
-  );
+    setActiveSection(currentSectionId);
+  });
 
   useEffect(() => {
     const scrollable = scrollingElement?.current ?? window;
